Implemented rudimentary predicitve forecast for Forecast tab.  Lapsed cards are assumed to never relapse within the timeframe.
* Stacking and colors are the same as in past-performance (Review Count), which makes
for nice comparisons.

* The forecast data represented is a count of card content, NOT of card views.  For
example, you will view a card you are learning usually 2x the day you first see it.
This card contributes 1 to my forecast statistic, and in the blue section.  This is
consistent with the previous construction of the Forecast statistic.  The code to
turn Forecast into the card views measurement is present, but I have commented it
in/out.

* As far as putting it into another pane: I have not done this because I'm not
familiar with your project's strings system, how to add panes, and there are ominous
words about breaking everything concerning the string system on the dev page, so I
possess the fear ;)

HARDCODES:
* The lapse rate is hardcoded at 10%
* The interval is hardcoded at base 2
* New cards is hardcoded at 20